+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 - Add complete husky config @sneridagh
 - Add COC.md file @timo
+- Add Cypress tests for blocks @timo @rodrigo @jakahl
 
 ## 4.0.0-alpha.5 (2019-09-28)